+++ eras/conway/impl/test/Main.hs
@@ -3,7 +3,6 @@
 module Main where
 
 import Cardano.Ledger.Conway (Conway)
-import Cardano.Ledger.Core (PParams)
 import qualified Test.Cardano.Ledger.Alonzo.Binary.CostModelsSpec as CostModelsSpec
 import Test.Cardano.Ledger.Common
 import qualified Test.Cardano.Ledger.Conway.Binary.CddlSpec as Cddl
@@ -13,7 +12,7 @@ import qualified Test.Cardano.Ledger.Conway.DRepRatifySpec as DRepRatify
 import qualified Test.Cardano.Ledger.Conway.GenesisSpec as Genesis
 import qualified Test.Cardano.Ledger.Conway.GovActionReorderSpec as GovActionReorder
 import qualified Test.Cardano.Ledger.Conway.Imp as ConwayImp
-import qualified Test.Cardano.Ledger.Core.JSON as JSON
+import Test.Cardano.Ledger.Core.JSON (roundTripJsonEraSpec)
 import qualified Test.Cardano.Ledger.Shelley.Imp as ShelleyImp
 
 main :: IO ()
@@ -26,8 +25,7 @@ main =
       CommitteeRatify.spec
       Genesis.spec
       GovActionReorder.spec
-      xdescribe "JSON" $ do
-        JSON.roundTripEraSpec @(PParams Conway)
+      roundTripJsonEraSpec @Conway
       describe "Imp" $ do
         ConwayImp.spec @Conway
         ShelleyImp.spec @Conway