+++ eras/shelley/impl/test/Main.hs
@@ -4,6 +4,7 @@ module Main where
 
 import Cardano.Ledger.Shelley (Shelley)
 import Test.Cardano.Ledger.Common
+import Test.Cardano.Ledger.Core.JSON (roundTripJsonEraSpec)
 import qualified Test.Cardano.Ledger.Shelley.Binary.CddlSpec as Cddl
 import qualified Test.Cardano.Ledger.Shelley.BinarySpec as Binary
 import qualified Test.Cardano.Ledger.Shelley.Imp as Imp
@@ -15,3 +16,4 @@ main =
       Binary.spec
       Cddl.spec
       Imp.spec @Shelley
+      roundTripJsonEraSpec @Shelley