+++ libs/cardano-ledger-core/test/Test/Cardano/Ledger/PlutusSpec.hs
@@ -17,12 +17,11 @@ spec :: Spec
 spec = do
   costModelsSpec
 
-
 costModelsSpec :: Spec
 costModelsSpec =
   describe "CostModels" $ do
-    prop "flattenCostModel . mkCostModelsLenient" $ \valid unknown -> do
-      let cms1 = flattenCostModel valid <> unknown
-          cms2 = unknown <> flattenCostModel valid
-      flattenCostModel (mkCostModelsLenient cms1) `shouldBe` cms1
-      flattenCostModel (mkCostModelsLenient cms2) `shouldBe` cms2
+    prop "flattenCostModels . mkCostModelsLenient" $ \valid unknown -> do
+      let cms1 = flattenCostModels valid <> unknown
+          cms2 = unknown <> flattenCostModels valid
+      flattenCostModels (mkCostModelsLenient cms1) `shouldBe` cms1
+      flattenCostModels (mkCostModelsLenient cms2) `shouldBe` cms2