+++ eras/babbage/impl/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.6.0.0
 
+* `FromJSON` instance for `PParamsUPdate`: #3949
+* Fixed JSON serialization of `PParams` to include the `protocolVersion`: #3953
 * Stop exporting all of the internal `hkd*` functions and `PParamsHKD` from
   `Cardano.Ledger.Babbage.Core`.
 * Stop exporting `AllegraEraTxBody`, `AlonzoEraTxBody`,`AlonzoTxBody`,
@@ -22,7 +24,6 @@
 * Remove `babbageScriptPrefixTag`
 * Moved `ToExpr` instances out of the main library and into the testlib.
 * Change the type of `ppEMaxL`
-* Fixed JSON serialization of `PParams` to include the `protocolVersion`.
 
 ## 1.5.1.0